Added correct sqlite datasource and fixed services and repositories

+++ Domain/Models/TaxReceipt.cs
@@ -3,7 +3,7 @@ namespace Domain.Models;
 public class TaxReceipt
 {
     public int Id { get; set; }
-    public int TaxPayerId { get; set; }
+    public long TaxPayerId { get; set; }
     public string Ncf { get; set; } = string.Empty;
     public decimal Amount { get; set; }
     public decimal Itbis18 { get; set; }
