+++ Infrastructure/interfaces/Services/ITaxReceiptService.cs
@@ -5,5 +5,5 @@ namespace Infrastructure.interfaces.Services;
 public interface ITaxReceiptService
 {
     Task<IEnumerable<TaxReceipt>> GetAllTaxReceipts();
-    Task<IEnumerable<TaxReceipt>> GetAllTaxReceiptsByTaxpayerId(int taxpayerId);
+    Task<IEnumerable<TaxReceipt>> GetAllTaxReceiptsByTaxpayerId(long taxpayerId);
 }
